# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MovieHost API",
    "description": "Media library, playback progress, and viewer profiles.",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/profile": {
      "post": {
        "tags": [
          "profiles"
        ],
        "summary": "Create a profile with a generated `UserN` username.",
        "operationId": "handle_post_profile",
        "responses": {
          "201": {
            "description": "Profile created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/profile/{id}": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "Fetch one profile together with every watch state it owns.",
        "operationId": "handle_get_profile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Profile id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Profile and watch states",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "profiles"
        ],
        "summary": "Rename a profile.",
        "operationId": "handle_put_profile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Profile id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Profile renamed"
          },
          "400": {
            "description": "Malformed id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "profiles"
        ],
        "summary": "Delete a profile. Its watch states cascade.",
        "operationId": "handle_delete_profile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Profile id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Profile deleted"
          },
          "400": {
            "description": "Malformed id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No such profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/profiles": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "List every profile, without watch states.",
        "operationId": "handle_get_profiles",
        "responses": {
          "200": {
            "description": "All profiles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Profile"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "type": "object",
        "description": "JSON body returned for every error response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable description of what went wrong.",
            "example": "Not found"
          }
        }
      },
...
      "Profile": {
        "type": "object",
        "description": "A viewer profile, without its watch states.",
        "required": [
          "id",
          "username"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "ProfileResponse": {
        "type": "object",
        "description": "A single profile plus every watch state it owns, keyed by media path.",
        "required": [
          "id",
          "username",
          "watch_states"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string"
          },
          "watch_states": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/WatchState"
            },
            "propertyNames": {
              "type": "string"
            }
          }
        }
      },
      "ProfileUpdate": {
        "type": "object",
        "description": "Request body for renaming a profile.",
        "required": [
          "username"
        ],
        "properties": {
          "username": {
            "type": "string"
          }
        }
      },
...
      "WatchState": {
        "type": "object",
        "description": "How far a profile got through one piece of media. Doubles as the request\nbody for `PUT /api/profile/{id}/watch_state`.",
        "required": [
          "movie_path",
          "last_position",
          "finished"
        ],
        "properties": {
          "finished": {
            "type": "boolean"
          },
          "last_position": {
            "type": "number",
            "format": "float",
            "description": "Playback position in seconds."
          },
          "movie_path": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}